rbd: reboot system of IO times out
This is just a hacky patch which can be used for experimental
AA. The idea is that we must have cleaned up running commands
before the initiator has failed it and retried on another
path and possibly sent new WRITEs to the same objects/sectors.

We have to 2 places where we need to detect network issues:

1. The iscsi connection.
2. The ceph connection.

For #1, the nopin_response_timeout and nopin_timeout must be set
low enough to be able to detect the iscsi connection has been dropped
and when this happens the target layer should inform tcmu-runner
to handle running commands [this is not done. Put link to upstream patch
here when sent].

For #2, we need to make sure ceph requests have either reached the
OSD and been queued so requests started after will execute behind them,
or that requests in flight are failed. To handle the case where a
ceph request is on the network but not yet on the OSD, this patch
just kills the system. We are looking into a nicer cleanup of
commands, but that is no where near down.

Note that even if the tcp connection is just dropped and an immediate
RST is sent (in our case it is just dropped), there may be many hops
between the rbd client and OSD and a tcp packet containing a ceph
request could be stuck in a switch which might make its way to the
OSD after the initiator has done failover. This patch does not handle
this case and assumes network timers

Signed-off-by: Mike Christie <mchristi@redhat.com>