rbd: cleanup blacklist
The ceph blacklist can grow so long it ends up causing cluster issues
like slow/failed READ/WRITE IO. This patch assumes we are the only ceph
bl user on the system and will cleanup entries when we reopen the
device.

TODO: We need a way to get the client nonce used, in case there are
other bl users on the same host.